Big update

+++ src/main/java/NetflixRatings.java
@@ -4,7 +4,7 @@ import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.fs.Path;
 
@@ -30,7 +30,7 @@ public class NetflixRatings extends Configured implements Tool {
 	    job.setReducerClass(MaxFilmRatingReducer.class);
 
 	    job.setOutputKeyClass(Text.class);
-	    job.setOutputValueClass(IntWritable.class);
+	    job.setOutputValueClass(DoubleWritable.class);
 	    
 	    return job.waitForCompletion(true) ? 0 : 1;
 	}
